+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 0.2.0 / 2017-09-30
 
 * [FEATURE] Add support + instructions for Circle CI 2.0
+* [FEATURE] Loosen codeclimate-test-reporter version dependency to allow for 1.x series
 
 # 0.1.2 / 2016-04-04